+++ package.json
@@ -16,7 +16,8 @@
     "next": "13.5.5",
     "pixi.js": "^7.1.4",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "tailwindcss-classnames": "^3.1.0"
   },
   "devDependencies": {
     "@types/node": "^20",
