+++ pages/globals.css
@@ -25,3 +25,10 @@ body {
     )
     rgb(var(--background-start-rgb));
 }
+
+a:hover {
+  color: blue;
+  text-decoration: underline;
+}
+
+
