+++ pages/index.module.css
@@ -2,15 +2,42 @@
     position: relative;
 }
 
-.blockNumber {
+.blockNumber,
+.chainName {
     position: absolute;
-    left: 0px;
-    top: 0px;
     font-family: monospace, monospace;
-    font-size: 40px;
+    font-size: 20px;
     padding-left: 25px;
+    padding-right: 25px;
     padding-top: 20px;
     /* border-bottom: solid; */
     color: #6C996B;    
+}
+
+.blockNumber {
+    right: 0px;
+    top: -20px;
+}
+
+.chainName {
+    left: 0px;
+    top: -20px;
+}
+
+
+
+.topContainer {
+    position: absolute;
+    top: 0px;
+    width: 100%;
+    height: 80px;
+    display: table;
+}
+
+
+.title {
+    display: table-cell;
+    text-align: center; 
+    vertical-align: bottom;
 }
