+++ tailwind.config.ts
@@ -14,6 +14,9 @@ const config: Config = {
           'conic-gradient(from 180deg at 50% 50%, var(--tw-gradient-stops))',
       },
     },
+    colors: {
+        'title-orange': '#E1952B',
+    }
   },
   plugins: [],
 }
